Extend BPF programs to check max. nb of tail calls
This commit extends kern.c to update an array map for the 34th (key=0)
and 35th (key=1) programs in the chain. As can be seen below, only the
value for key=0 is updated to 1, demonstrating that program 34th was
executed but not program 35th. The chain is therefore limited to 34
programs, that is 33 tail calls, on the x86-64 JIT at least.

    $ ./run.sh 34 1
    [sudo] password for paul:
    [ 201, 345, 494, 662, 700, 737, 884, 999, 1071, 1406, 1231, 1589, 1438, 1548, 2026, 2245, 2056, 2058, 1893, 1987, 2309, 2241, 2507, 2384, 2700, 2422, 2956, 2760, 2852, 3926, 3225, 3674, 3639, 3796, 3617, ]
    $ sudo bpftool map dump name count
    key: 00 00 00 00  value: 01 00 00 00
    key: 01 00 00 00  value: 00 00 00 00
    Found 2 elements
    $ uname -r
    5.5.19-050519-generic
    $ sysctl net.core.bpf_jit_enable
    net.core.bpf_jit_enable = 1

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>